+++ src/v1/Archetypes/UtilityTypes/Definitely.ts
@@ -32,18 +32,16 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-export * from "./assignOptionalFields";
-export * from "./assignOptionalFieldsUsingTransformers";
-export * from "./getAllMethodNames";
-export * from "./getAllMethods";
-export * from "./getMissingMethodNames";
-export * from "./getPublicMethods";
-export * from "./getPublicMethodNames";
-export * from "./isAttributeName";
-export * from "./isGetterName";
-export * from "./isMethodName";
-export * from "./isObject";
-export * from "./mustBeObject";
-export * from "./validateObject";
-export * from "./validateObjectHasAllMethodsCalled";
-export * from "./Filters";
+/**
+ * `Definitely` is a utility type. Use it to remove `undefined` from
+ * the union of possible types of `T`.
+ *
+ * It's a gentler alternative to `NonNullable`, and the exact opposite
+ * of {@link Maybe}
+ *
+ * @template T
+ * A source data type.
+ *
+ * @category UtilityTypes
+ */
+export type Definitely<T> = Exclude<T, undefined>;