+++ src/v1/Archetypes/UtilityTypes/TransformedAttributes.ts
@@ -31,19 +31,33 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
+import { AttributeTransformers } from "./AttributeTransformers";
 
-export * from "./assignOptionalFields";
-export * from "./assignOptionalFieldsUsingTransformers";
-export * from "./getAllMethodNames";
-export * from "./getAllMethods";
-export * from "./getMissingMethodNames";
-export * from "./getPublicMethods";
-export * from "./getPublicMethodNames";
-export * from "./isAttributeName";
-export * from "./isGetterName";
-export * from "./isMethodName";
-export * from "./isObject";
-export * from "./mustBeObject";
-export * from "./validateObject";
-export * from "./validateObjectHasAllMethodsCalled";
-export * from "./Filters";
+
+/**
+ * `TransformedAttributes` is a mapped type. It contains a list of fields
+ * taken from `T`, and the type of these fields after the transformers
+ * in `AT` have been applied.
+ *
+ * ie, given:
+ * - two types `S` (source type) and `T` (target type),
+ * - and a set of {@link AttributeTransformers} `AT<S>` to convert `S`
+ *   into `T`,
+ * it creates a type that contains the fields of 'T' that 'AT' produces.
+ *
+ * Most importantly, it does it without you having to provide 'T' as a
+ * type parameter.
+ *
+ * @template S
+ * the type that the {@link AttributeTransformers} are applied to
+ * @template AT
+ * the list of fields to be created, and the functions that do the
+ * creating
+ *
+ * @category UtilityTypes
+ */
+export type TransformedAttributes<S extends object, AT extends AttributeTransformers<S>> = {
+    [K in keyof AT]: K extends keyof S ?
+        AT[K] extends (...args: any[]) => infer R ? R : void
+        : never
+}